pipeline-registry/wonderware_to_clickhouse/v1/514-labs/python/default/requirements.txt
# Core
moose-cli>=0.6.230
moose-lib>=0.6.230
pydantic>=2.11.0

# Database
sqlalchemy>=2.0.0
python-tds>=1.16.0
clickhouse-connect>=0.7.0

# Utilities
tenacity>=9.0.0
python-dotenv>=1.0.0
redis>=5.0.0

# Testing
pytest>=7.0.0
